--- src/sider/eu_eøs/stegKomponenter/vurderingArbeidsmonster/landlinje/landlinje.less
@@ -1,19 +1,0 @@
-.vurderingArbeidsmonster__landlinje {
-  display: flex;
-  align-items: flex-end;
-  justify-content: space-between;
-  padding: 0.5rem 10% 0.5rem 0;
-  border-bottom: 1px solid black;
-
-  :first-child {
-    width: 40%;
-  }
-
-  &__marginaltArbeidCheckbox {
-    width: 20%;
-  }
-
-  * {
-    margin-bottom: 0; // Fjerner margin på Nav.Checkbox for å passe inn i rad
-  }
-}
